fix: renamed struct variables to match new naming scheme
Also fixed any code that was dependent on the structs. Created some more testing code in Firebase example, with an example of how to iterate through all anime objects in a dictionary. Deleted references/code to the old description and recap structs, and cleaned up code in AnimeFirebaseData and FirebaseExample overall