feat: implemented Firebase into most views
Firebase was used with the AnimeTracker and Categories views. It still needs to be implemented to the Schedule views, which will be done at a later date. Firebase is implemented very badly, and does not use an @Environment, which causes the app to crash when trying to open the actual app and not just individual views. In the views that have implemented Firebase, notably the favoruite show feature isn't implemented with the Firebase version, which requires the original json data to still be linked just to build the views. The plan is to 1. implement favourites properly so we can fully transition to using Firebase in all views, and 2. to use Environments to properly manage resources in the app.