feat: implemented animeDataFB to use @EnvironmentObject
Currently, the favorites bug is still present but I have found the solution and just need to implement it!

Furthur detail about future favorites bug fix: So, I was not referencing the anime object from the @EnvironmentObject, which meant that it would not update the changes. So, I just need to recode/reformat all of the views which are structured like that (where I have to pass in anime objects, etc.) to just pass in the anime doc ID instead, and fetch the desired value via @EnvironmentObject from inside that class. Otherwise, this problem is solved!