test: implemented retrieval of data directly into codable struct
We used to manually add items into dictionary spots, and used to retrieve items by their key. Now, they are treated as structs and we are able to easily decode/encode our data! This is not yet fully implemented for all data, but is still in testing and will be implemented soon