fix: forces toggles to switch to "off" state when no user is authenticated
Previously if a user was authenticated, enabled the toggle, and signed out, the toggles would be disabled but would still stay in the "enabled toggle" state. This was bad as 1. it doesn't make sense, and 2. in the List view, it would "filter" by favorites even though there were no favorites, so it would show nothing. This code fixes that and forces the toggles to turn off